part4 exercise 4.21 commit

+++ part4/blog/controllers/blogs.js
@@ -1,5 +1,5 @@
 const blogsRouter = require('express').Router()
-const { request } = require('../app')
+//const { request } = require('../app')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
@@ -10,7 +10,7 @@ blogsRouter.get('/', async (request, response) => {
         .find({})
         .populate('user',{ username:1, name:1 })
 
-    response.json(blogs)
+    return response.json(blogs)
 
     /*
     Blog
@@ -44,10 +44,10 @@ blogsRouter.post('/', async (request, response) => {
         const updatedUser = await User.findByIdAndUpdate(user._id, { blogs: user.blogs.concat([blog.id]) }, { new: true })
         console.log(updatedUser)
 
-        response.status(201).json(result)
+        return response.status(201).json(result)
     }
 
-    response.status(400).json({ error: 'Missing blog title or url' })
+    return response.status(400).json({ error: 'Missing blog title or url' })
 
     /*
     blog
@@ -60,12 +60,30 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const blog = await Blog.findByIdAndDelete(request.params.id)
-    if (blog)
-    {
-        response.status(204).end()
+
+    // verify token
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'invalid token' })
     }
-    response.status(404).end()
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog)
+    {
+        return response.status(404).json({ error: 'invalid blog id' })
+    }
+
+    // check if the blog creator and token user are same
+    if (blog.user.toString() === decodedToken.id.toString())
+    {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }
+    else
+    {
+        return response.status(401).json({ error: 'Deletion on object not allowed.' })
+    }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -77,10 +95,10 @@ blogsRouter.put('/:id', async (request, response) => {
 
     const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     if (updatedBlog) {
-        response.json(updatedBlog)
+        return response.json(updatedBlog)
     }
 
-    response.status(404).end()
+    return response.status(404).end()
 })
 
 module.exports = blogsRouter
